Move Settings and TaskTree into routes
