+++ src/routes/[user]/components/ListsArea/TodoList.svelte
@@ -1,6 +1,6 @@
 <script>
-  import Dropzone from '$lib/TaskTree/Dropzone.svelte'
-  import RecursiveTask from '$lib/TaskTree/RecursiveTask.svelte'
+  import Dropzone from '../../components/TaskTree/Dropzone.svelte'
+  import RecursiveTask from '../../components/TaskTree/RecursiveTask.svelte'
   import FormField from '$lib/Reusable/FormField.svelte'
   import { getRandomID } from '/src/utils/core.js'
   import { HEIGHTS } from '/src/utils/constants.js'
