+++ src/routes/[user]/components/Settings/ColorSettings.svelte
