+++ src/routes/[user]/components/TaskTree/Dropzone.svelte
@@ -17,7 +17,7 @@
 <script>
   import { user, activeDragItem } from '/src/store'
   import { increment, writeBatch, doc } from 'firebase/firestore'
-  import { db } from '../../db/init'
+  import { db } from '/src/db/init'
   import Task from '/src/db/models/Task.js'
   import { HEIGHTS } from '/src/utils/constants.js'
   import { listTreeNodes } from '/src/db/models/treeISOs.js'
